requirements.txt
Django==1.11.8
django-bootstrap3==9.0.0
django-markdown-deux==1.0.5
djangorestframework==3.7.7
markdown2==2.3.4
pytz==2017.2
